BAEL-1029 - deep20jain@gmail.com - Addressing review comment (#2294)
* Adding test classes for java atomic variables

* Updating counter with atomic integer

* Adding reason for ignoring test

* Removing ignore annotation and moving test to manual test

* Removing counter test
